test: created tests for all classes

+++ src/main/kotlin/org/pecasonline/features/subscription/SubscriptionRepository.kt
@@ -2,4 +2,4 @@ package org.pecasonline.features.subscription
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface SubscriptionRepository: JpaRepository<Subscription, Int>
+interface SubscriptionRepository: JpaRepository<Subscription, Long>
